Refactoring

+++ src/main/java/org/tdl/vireo/model/repo/custom/DepositLocationRepoCustom.java
@@ -6,8 +6,8 @@ public interface DepositLocationRepoCustom {
 
     public DepositLocation create(String name);
     
-    public void reorder(Integer src, Integer dest);
+    public void reorder(Long src, Long dest);
     
-    public void remove(Integer index);
+    public void remove(Long index);
 
 }
