+++ src/main/java/org/tdl/vireo/model/repo/custom/LanguageRepoCustom.java
@@ -6,10 +6,10 @@ public interface LanguageRepoCustom {
 
     public Language create(String name);
         
-    public void reorder(Integer src, Integer dest);
+    public void reorder(Long src, Long dest);
     
     public void sort(String column);
     
-    public void remove(Integer index);
+    public void remove(Long index);
 
 }
